Move granules api functions to their own file
