+++ example/spec/parallel/ingest/ingestFromPdrSpec.js
@@ -34,6 +34,7 @@ const {
   buildAndExecuteWorkflow,
   cleanupProviders,
   cleanupCollections,
+  granulesApi: granulesApiTestUtils,
   LambdaStep,
   waitForCompletedExecution
 } = require('@cumulus/integration-tests');
@@ -222,7 +223,7 @@ describe('Ingesting from PDR', () => {
         await Promise.all(queueGranulesOutput.payload.running.map(async (arn) => {
           await waitForCompletedExecution(arn);
         }));
-        await apiTestUtils.deleteGranule({
+        await granulesApiTestUtils.deleteGranule({
           prefix: config.stackName,
           granuleId: parseLambdaOutput.payload.granules[0].granuleId
         });
@@ -307,7 +308,7 @@ describe('Ingesting from PDR', () => {
           // delete ingested granule(s)
           await Promise.all(
             finalOutput.payload.granules.map((g) =>
-              apiTestUtils.deleteGranule({
+              granulesApiTestUtils.deleteGranule({
                 prefix: config.stackName,
                 granuleId: g.granuleId
               }))
