+++ example/spec/parallel/syncGranule/SyncGranuleDuplicateHandlingSpec.js
@@ -14,6 +14,7 @@ const {
   buildAndExecuteWorkflow,
   cleanupCollections,
   cleanupProviders,
+  granulesApi: granulesApiTestUtils,
   LambdaStep
 } = require('@cumulus/integration-tests');
 const {
@@ -60,8 +61,6 @@ describe('When the Sync Granule workflow is configured', () => {
   const provider = { id: `s3_provider${testSuffix}` };
   const newCollectionId = constructCollectionId(collection.name, collection.version);
 
-  const fileStagingDir = 'custom-staging-dir';
-
   let inputPayload;
   let expectedPayload;
   let workflowExecution;
@@ -94,7 +93,7 @@ describe('When the Sync Granule workflow is configured', () => {
       deleteFolder(config.bucket, testDataFolder),
       cleanupCollections(config.stackName, config.bucket, collectionsDir, testSuffix),
       cleanupProviders(config.stackName, config.bucket, providersDir, testSuffix),
-      apiTestUtils.deleteGranule({
+      granulesApiTestUtils.deleteGranule({
         prefix: config.stackName,
         granuleId: inputPayload.granules[0].granuleId
       })
@@ -193,7 +192,7 @@ describe('When the Sync Granule workflow is configured', () => {
       });
 
       it('captures both files', async () => {
-        const granuleResponse = await apiTestUtils.getGranule({
+        const granuleResponse = await granulesApiTestUtils.getGranule({
           prefix: config.stackName,
           granuleId: inputPayload.granules[0].granuleId
         });
@@ -237,7 +236,7 @@ describe('When the Sync Granule workflow is configured', () => {
       });
 
       it('captures all files', async () => {
-        const granuleResponse = await apiTestUtils.getGranule({
+        const granuleResponse = await granulesApiTestUtils.getGranule({
           prefix: config.stackName,
           granuleId: inputPayload.granules[0].granuleId
         });
