+++ packages/integration-tests/api/api.js
@@ -53,28 +53,6 @@ async function callCumulusApi({ prefix, functionName, payload: userPayload }) {
   return JSON.parse(apiOutput.Payload);
 }
 
-/**
- * GET /granules/{granuleName}
- *
- * @param {Object} params - params
- * @param {string} params.prefix - the prefix configured for the stack
- * @param {string} params.granuleId - a granule ID
- * @returns {Promise<Object>} - the granule fetched by the API
- */
-async function getGranule({ prefix, granuleId }) {
-  return callCumulusApi({
-    prefix: prefix,
-    functionName: 'ApiGranulesDefault',
-    payload: {
-      httpMethod: 'GET',
-      resource: '/granules/{granuleName}',
-      path: `/granules/${granuleId}`,
-      pathParameters: {
-        granuleName: granuleId
-      }
-    }
-  });
-}
 
 /**
  * GET /asyncOperations/{id}
@@ -115,109 +93,6 @@ function postBulkDelete({ prefix, granuleIds }) {
       path: '/bulkDelete',
       pathParameters: {},
       body: JSON.stringify({ granuleIds })
-    }
-  });
-}
-
-/**
- * Reingest a granule from the Cumulus API
- *
- * @param {Object} params - params
- * @param {string} params.prefix - the prefix configured for the stack
- * @param {string} params.granuleId - a granule ID
- * @returns {Promise<Object>} - the granule fetched by the API
- */
-async function reingestGranule({ prefix, granuleId }) {
-  return callCumulusApi({
-    prefix: prefix,
-    functionName: 'ApiGranulesDefault',
-    payload: {
-      httpMethod: 'PUT',
-      resource: '/granules/{granuleName}',
-      path: `/granules/${granuleId}`,
-      pathParameters: {
-        granuleName: granuleId
-      },
-      body: JSON.stringify({ action: 'reingest' })
-    }
-  });
-}
-
-/**
- * Removes a granule from CMR via the Cumulus API
- *
- * @param {Object} params - params
- * @param {string} params.prefix - the prefix configured for the stack
- * @param {string} params.granuleId - a granule ID
- * @returns {Promise<Object>} - the granule fetched by the API
- */
-async function removeFromCMR({ prefix, granuleId }) {
-  const payload = await callCumulusApi({
-    prefix: prefix,
-    functionName: 'ApiGranulesDefault',
-    payload: {
-      httpMethod: 'PUT',
-      resource: '/granules/{granuleName}',
-      path: `/granules/${granuleId}`,
-      pathParameters: {
-        granuleName: granuleId
-      },
-      body: JSON.stringify({ action: 'removeFromCmr' })
-    }
-  });
-
-  try {
-    return payload;
-  }
-  catch (error) {
-    console.log(`Error parsing JSON response removing granule ${granuleId} from CMR: ${JSON.stringify(payload)}`);
-    throw error;
-  }
-}
-/**
- * Run a workflow with the given granule as the payload
- *
- * @param {Object} params - params
- * @param {string} params.prefix - the prefix configured for the stack
- * @param {string} params.granuleId - a granule ID
- * @param {string} params.workflow - workflow to be run with given granule
- * @returns {Promise<Object>} - the granule fetched by the API
- */
-async function applyWorkflow({ prefix, granuleId, workflow }) {
-  return callCumulusApi({
-    prefix: prefix,
-    functionName: 'ApiGranulesDefault',
-    payload: {
-      httpMethod: 'PUT',
-      resource: '/granules/{granuleName}',
-      path: `/granules/${granuleId}`,
-      pathParameters: {
-        granuleName: granuleId
-      },
-      body: JSON.stringify({ action: 'applyWorkflow', workflow })
-    }
-  });
-}
-
-/**
- * Delete a granule from Cumulus via the API
- *
- * @param {Object} params - params
- * @param {string} params.prefix - the prefix configured for the stack
- * @param {string} params.granuleId - a granule ID
- * @returns {Promise<Object>} - the delete confirmation from the API
- */
-async function deleteGranule({ prefix, granuleId }) {
-  return callCumulusApi({
-    prefix: prefix,
-    functionName: 'ApiGranulesDefault',
-    payload: {
-      httpMethod: 'DELETE',
-      resource: '/granules/{granuleName}',
-      path: `/granules/${granuleId}`,
-      pathParameters: {
-        granuleName: granuleId
-      }
     }
   });
 }
@@ -559,10 +434,8 @@ async function updateProvider({ prefix, provider, updateParams }) {
 }
 
 module.exports = {
-  applyWorkflow,
   callCumulusApi,
   getAsyncOperation,
-  deleteGranule,
   deletePdr,
   getExecution,
   getExecutionLogs,
@@ -575,11 +448,8 @@ module.exports = {
   getWorkflow,
   getProvider,
   getCollection,
-  getGranule,
   getLogs,
   postBulkDelete,
-  reingestGranule,
-  removeFromCMR,
   updateCollection,
   updateProvider
 };
