+++ packages/integration-tests/index.js
@@ -29,6 +29,7 @@ const {
 const sfnStep = require('./sfnStep');
 const api = require('./api/api');
 const rulesApi = require('./api/rules');
+const granulesApi = require('./api/granules');
 const cmr = require('./cmr.js');
 const lambda = require('./lambda');
 const granule = require('./granule.js');
@@ -691,6 +692,7 @@ async function waitForTestExecutionStart(
 module.exports = {
   api,
   rulesApi,
+  granulesApi,
   buildWorkflow,
   testWorkflow,
   executeWorkflow,
